Dialog Fragment used for Filters
For updating the FilterDialog with the date choosed by the user in DatePickerDialog I followed the next strategy. In first place, the filter dialog it register itself as a Shared Preferences listener. In second place I created a new MyDatePickerDialog that extended from DatePickerDialog and overrided onDateChanged, so each time it was called I write the date in SharedPreferences. That way the Filter Dialog was called and then I could update the EditText where the date choosed was represented.